+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - New short-circuiting function `mapping.heuristic_functions.smurf_columns()`.
+- New function `dio.register_io()`, allowing users to create their own custom IO implementations.
 
 ### Changed
 - Lower cache hit default log level from INFO to DEBUG.